requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
tensorflow
opencv-python
flask
xgboost
mlflow
PyYAML
joblib
shap
albumentations